refactor: treat required parameters
