+++ src/config-file/parameters/display-name.ts
@@ -3,6 +3,7 @@ import type { ConfigFileParameter } from "@/types/config-file";
 import { t } from "vscode-ext-localisation";
 
 export const DISPLAY_NAME = {
+	required: true,
 	validation(keys, value, line, diagnostics, document) {
 		if (value.length > 32) {
 			diagnostics.push(
