+++ src/config-file/parameters/start.ts
@@ -3,6 +3,7 @@ import type { ConfigFileParameter } from "@/types/config-file";
 import { t } from "vscode-ext-localisation";
 
 export const START = {
+	required: false,
 	validation(keys, value, line, diagnostics, document) {
 		if (value.length > 128) {
 			diagnostics.push(
